Move kubectl cp back to k/k
